Adding Game's logic
Update on specific call to arguments: it was a user prompt and it
is now a proper argument when launching the programs.

Update on class Server, PostHandler, PingHandler, Launcher and client
while using the new Parser.

Adding class Attack, BoardCells, BattleShips, Sea, State and
GetHandler for game's board game and behaviours between ships